Engine1 v. 0.4

+++ engines/First OGL Engine/src/NormalRenderer.cpp
@@ -18,10 +18,10 @@ namespace OGL::E1 {
         m_shader.setUniformVec3("viewerPos", camera->getPos());
 
         dirLights &sceneDirLights = scene.getDirLights();
-        m_shader.setUniformInt("numDirLights", sceneDirLights.size());
+        m_shader.setUniformInt("numDirLights", static_cast<int>(sceneDirLights.size()));
         for (size_t i = 0; i < sceneDirLights.size(); ++i) {
-            auto &[light, pShadowMap] = sceneDirLights[i];
-            light.loadInShader(m_shader, i);
+            auto& [light, pShadowMap] = sceneDirLights[i];
+            light.loadInShader(m_shader, static_cast<int>(i));
             if (!pShadowMap) {
                 continue;
             }
@@ -31,10 +31,10 @@ namespace OGL::E1 {
         }
 
         pointLights &scenePointLights = scene.getPointLights();
-        m_shader.setUniformInt("numPointLights", scenePointLights.size());
+        m_shader.setUniformInt("numPointLights", static_cast<int>(scenePointLights.size()));
         for (size_t i = 0; i < scenePointLights.size(); ++i) {
-            auto &[light, pShadowCubemap] = scenePointLights[i];
-            light.loadInShader(m_shader, i);
+            auto& [light, pShadowCubemap] = scenePointLights[i];
+            light.loadInShader(m_shader, static_cast<int>(i));
             if (!pShadowCubemap) {
                 continue;
             }
@@ -44,10 +44,10 @@ namespace OGL::E1 {
         }
 
         spotLights &sceneSpotLights = scene.getSpotLights();
-        m_shader.setUniformInt("numSpotLights", sceneSpotLights.size());
+        m_shader.setUniformInt("numSpotLights", static_cast<int>(sceneSpotLights.size()));
         for (size_t i = 0; i < sceneSpotLights.size(); ++i) {
-            auto &[light, pShadowMap] = sceneSpotLights[i];
-            light.loadInShader(m_shader, i);
+            auto& [light, pShadowMap] = sceneSpotLights[i];
+            light.loadInShader(m_shader, static_cast<int>(i));
             if (!pShadowMap) {
                 continue;
             }
